add Firebase class
